Return fake NUMA cell memory when NUMA unavailable
When NUMA is not available on a host, effectively there is only 1
NUMA cell in the topology. However, numactl will raise an error in
this case, and zun-compute will fail to start. Handle this case
gracefully by falling back to returning a fake NUMA cell with the
existing CPU info obtained from `lscpu` and cell memory == the total
available memory on the node.