Add K8s driver for container requests
This adds a new container driver for Zun, which schedules containers
on a K8s cluster. Currently Zun only has one driver, which interacts
with the Docker daemon and schedules containers locally on a host.
The K8s driver differs in that it's not scheduling anything locally;
it works more like Nova's Ironic driver, which sets up a "proxy" compute
node, which acts as the frontend to Ironic. Similarly, the K8s driver
should be deployed on a control node and not on a compute node, and it
creates a single "node" in Zun's DB, which will have resources equal to
the total resources available on the entire K8s cluster.

It is possible to run this driver and the Docker driver in parallel.
The K8s compute agent will get a label "container_driver" label, which
can be used with the new DriverFilter to target only hosts using a
given container driver.

Not all operations are supported w/ the K8s driver, notably "commit"
and the archive get/put operations don't have clear analogues in K8s.
These will currently raise an exception.

Internally, a container request in Zun is translated into a K8s
Deployment resource with a single pod having a single container. This
does mean that if the pod is deleted somehow, it will be automatically
recreated. This can support something resembling "restart" in the Zun
lifecycle (where the pod is deleted and then recreated by the
Deployment), w/ some differences in behavior (the ephemeral file system
will be reset.)