Remove Cyborg integration
This removes the Cyborg integration from Zun entirely. Mostly, this is
due to the fact that in K8s we are not using the scheduling mechanism
to request resources and it will be more work to make this generic; any
device requests for k8s-backed containers should always get scheduled
to the zun-compute-k8s host (and be scheduled by k8s), rather than
having Zun schedule the container via information obtained from Placement.