Added toleration mechanism for worker node taint
Zun now takes a worker node taint into its conf and will proceed to set
a toleration for it in k8s deployments.

The taint key, value, and effect have to be specified as strings in the
site's config as

- k8s_worker_taint
  - key: taint key, by default set to "worker-node"
  - value: taint value, by default set to "true"
  - effect: taint effect, by default set to "NoSchedule"

under kolla/defaults.yml in chi-in-a-box:

- worker_taint:
  - key: {{ k8s_worker_taint.key | default("worder-node") }}
  - value: {{ k8s_worker_taint.value | default("true") }}
  - effect: {{ k8s_worker_taint.effect | default("NoSchedule") }}

the zun.conf options are set in the following way

- worker_taint_key: {{ worker_taint.key }}
- worker_taint_value: {{ worket_taint.value }}
- worker_taint_effect: {{ worker_taint.effect }}

Tolerating worker taint is enabled if the zun_enforce_worker_taint site config
option is set to True, by default, it is set to False.