fix sync when nameless images are present
periodic tasks were hanging on the same error as here https://review.opendev.org/c/starlingx/config/+/764098

kubernetes-client/python#895
If a container image contains no tag or digest, patch/list
node requests sent via python Kubernetes client will be
returned with exception because python Kubernetes client
deserializes the ContainerImage response from kube-apiserver
and it fails the validation due to the empty image name.

Implement this workaround to replace the V1ContainerImage.names
in the python Kubernetes client to bypass the "none image"
check because the error is not from kubernetes. If patching
a node with a new host label, we can see the label is
created successfully in Kubernetes.

This workaround should be removed if the proposed solutions
can be made in kubernetes or a workaround can be implemented
in containerd.
kubernetes/kubernetes#79018
containerd/containerd#4771