map correct fields to status_detail this time
zun's status detail field is only 50 chars, and intended for a short set
of summary messages. In k8s, "reason" fills a similar role, so map this
to status detail. The zun "status_reason" field is free-form text, so we
map the k8s "message" field.

In future, we could add additional information to status_detail, rather
than a 1-1 mapping.